requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
autopep8==1.6.0
Django==3.2
django-cors-headers==3.12.0
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==9.1.1
promise==2.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==1.7.1
pytz==2022.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tzdata==2022.1
pyparsing==3.0.9
pydot==1.4.2
django-extensions==3.1.5
graphene-file-upload==1.3.0
